Update processFile
This implements a new FileIsrTask that mangles the file to allow
passing to ProcessCcdTask.  This also makes ProcessFileTask just
a very thin wrapper that does little other than set config defaults.

Also move things to bin.src so that the shebang function can be run
on the command line tools.